gh-108303: Move `test_future` into its own subdir
