+++ Lib/test/test_future_import/badsyntax_future10.py
