+++ Lib/test/test_future_import/badsyntax_future3.py
