+++ Lib/test/test_future_import/badsyntax_future4.py
