+++ Lib/test/test_future_import/badsyntax_future6.py
