+++ Lib/test/test_future_import/badsyntax_future7.py
