+++ Makefile.pre.in
@@ -2160,6 +2160,7 @@ TESTSUBDIRS=	idlelib/idle_test \
 		test/test_dataclasses \
 		test/test_email \
 		test/test_email/data \
+		test/test_future_import \
 		test/test_import \
 		test/test_import/data \
 		test/test_import/data/circular_imports \
